requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
SQLAlchemy>=2.0.44
psycopg2-binary>=2.9.10
chromadb>=0.5.0
pydantic>=2.12.5
pydantic-settings>=2.12.0
python-multipart>=0.0.21
requests>=2.32.5
pypdf>=5.4.0
pdfplumber>=0.11.9
openai>=1.50.0
streamlit>=1.40.0
watchdog>=6.0.0
python-jose[cryptography]>=3.5.0
bcrypt>=4.2.0
python-dateutil>=2.9.0
rapidfuzz>=3.14.3
slowapi>=0.1.9
redis>=5.0.0
pandas>=2.1.4
streamlit-mic-recorder
httpx>=0.27.0
python-dotenv>=1.0.1
extra-streamlit-components>=0.1.71
alembic>=1.13.1
neo4j>=5.19.0
streamlit-agraph>=0.0.45
pytesseract>=0.3.10
pdf2image>=1.17.0
